improved the app package layout and added a logout button
