+++ main/routes/templates/routes/search.html
@@ -1,0 +1,20 @@
+{% extends "base.html" %}
+{% block title %}
+Recommendations
+{% endblock %}
+
+{% block content %}
+
+{% for song_data in data %}
+<img src="{{ song_data[3] }}" alt="Album Art" style="width:300px;height:300px;">
+<br>
+
+<button type="submit" name="song_chosen" formtarget="{{ url_for('main.index') }}">
+
+    {{ song_data [1] }} - {{ song_data [2] }}
+
+</button>
+<br><br>
+{% endfor %}
+
+{% endblock %}
